% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countTPM.R
\name{countTPM}
\alias{countTPM}
\title{countTPM}
\usage{
countTPM(rawdata, clean, x)
}
\arguments{
\item{clean}{whether to clean the data}

\item{x}{what method would you like to use to clean your data}

\item{DATA}{raw data}
}
\description{
Analyse RNA-seq data
}
