% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differenceANALYSIS.R
\name{differenceANALYSIS}
\alias{differenceANALYSIS}
\title{differenceANALYSIS}
\usage{
differenceANALYSIS(rawdata, group, omitNA)
}
\arguments{
\item{rawdata}{raw data}

\item{group}{experiment group}

\item{omitNA}{omitNAs}
}
\description{
Analyse RNA-seq data
}
