% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOWNSTREAManalysis.R
\name{DOWNSTREAManalysis}
\alias{DOWNSTREAManalysis}
\title{DOWNSTREAManalysis}
\usage{
DOWNSTREAManalysis(DATA, GROUP, GENE_TYPE, SPECIES)
}
\arguments{
\item{DATA}{raw data}

\item{GROUP}{experiment group}

\item{GENE_TYPE}{gene type annotated in raw data}

\item{SPECIES}{species}
}
\description{
Analyse RNA-seq data
}
