% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGGenrich.R
\name{KEGGenrich}
\alias{KEGGenrich}
\title{KEGGenrich}
\usage{
KEGGenrich(GENE_TO_ENRICH, SPECIES)
}
\arguments{
\item{GENE_TO_ENRICH}{gene list generated by IDchange}

\item{SPECIES}{specie}

\item{CHANGEtoTYPE}{type of gene}
}
\description{
Analyse RNA-seq data
}
