% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBAR.R
\name{plotBAR}
\alias{plotBAR}
\title{plotBAR}
\usage{
plotBAR(x, enrichTYPE)
}
\arguments{
\item{x}{GO or KEGG enrich analized data}

\item{enrichTYPE}{GO or KEGG enrich type}
}
\description{
Analyse RNA-seq data
}
