% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHEATMAP1.R
\name{plotHEATMAP1}
\alias{plotHEATMAP1}
\title{plotHEATMAP1}
\usage{
plotHEATMAP1(difference_analysised_DATA, HEATMAP_COL_GROUP)
}
\arguments{
\item{difference_analysised_DATA}{difference analysised data with differenceANALYSE}

\item{HEATMAP_COL_GROUP}{colgroup}
}
\description{
Analyse RNA-seq data
}
