% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHEATMAP2.R
\name{plotHEATMAP2}
\alias{plotHEATMAP2}
\title{plotHEATMAP2}
\usage{
plotHEATMAP2(difference_analysised_DATA, HEATMAP_COL_GROUP)
}
\arguments{
\item{difference_analysised_DATA}{difference analysised data with differenceANALYSE}

\item{HEATMAP_COL_GROUP}{colgroup}
}
\description{
Analyse RNA-seq data
}
