% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVOCANOL.R
\name{plotVOCANOL}
\alias{plotVOCANOL}
\title{plotVOCANOL}
\usage{
plotVOCANOL(differenceDATA, LOG2FC, PVALUE, TAQ, TITLE)
}
\arguments{
\item{differenceDATA}{analised data with differenceANALYSE}

\item{LOG2FC}{logFC cutoff}

\item{PVALUE}{Pvalue cutoff}

\item{TAQ}{add taqs in your interested genes}

\item{TITLE}{the title of the plot}
}
\description{
Analyse RNA-seq data
}
