% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOenrich.R
\name{GOenrich}
\alias{GOenrich}
\title{GOenrich}
\usage{
GOenrich(GENE_TO_ENRICH, SPECIES)
}
\arguments{
\item{GENE_TO_ENRICH}{gene list generated by IDchange}

\item{SPECIES}{specie}

\item{CHANGEtoTYPE}{type of gene}
}
\description{
Analyse RNA-seq data
}
