% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDchange.R
\name{IDchange}
\alias{IDchange}
\title{IDchange}
\usage{
IDchange(
  DATA,
  geneIDtoCHANGE,
  CHANGEfromTYPE,
  CHANGEtoTYPE,
  SPECIES,
  omitNA = TRUE,
  MERGEtoDATA = FALSE,
  drop = TRUE
)
}
\arguments{
\item{DATA}{raw data}

\item{geneIDtoCHANGE}{gene list, may from rawdata}

\item{CHANGEfromTYPE}{type of gene}

\item{CHANGEtoTYPE}{type of gene}

\item{SPECIES}{specie}

\item{omitNA}{omitNA}

\item{MERGEtoDATA}{do you want to merge chanegd geneid to the rawdata}
}
\description{
Analyse RNA-seq data
}
