% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqanalyseR.R
\docType{pacakge}
\name{seqanalyseR}
\alias{seqanalyseR}
\title{seqanalyseR : A package for RNA-seq}
\description{
A package to provide RNA-seq downstream analysis
}
\section{test Functions }{

RNA-seq analysis
}

